+++ lib/mongo.rb
@@ -33,7 +33,7 @@ require 'mongo/operation'
 require 'mongo/pool'
 require 'mongo/protocol'
 require 'mongo/server'
-require 'mongo/read_preference'
+require 'mongo/server_selector'
 require 'mongo/socket'
 require 'mongo/uri'
 require 'mongo/version'
