+++ lib/mongo/client.rb
@@ -163,7 +163,7 @@ module Mongo
     #
     # @since 2.0.0
     def read_preference
-      @read_preference ||= ReadPreference.get(options[:read] || {})
+      @read_preference ||= ServerSelector.get(options[:read] || {})
     end
 
     # Use the database with the provided name. This will switch the current
