+++ lib/mongo/collection.rb
@@ -32,7 +32,7 @@ module Mongo
     # @return [ Hash ] The collection options.
     attr_reader :options
 
-    # Get client, cluser and read preference from client.
+    # Get client, cluster, read preference, and write concern from client.
     def_delegators :database, :client, :cluster, :read_preference, :write_concern
 
     # Delegate to the cluster for the next primary.
