+++ lib/mongo/collection/view/aggregation.rb
@@ -98,7 +98,7 @@ module Mongo
             initial_query_op.execute(server.context)
           rescue Mongo::Operation::Aggregate::NeedPrimaryServer
             warn 'Rerouting the Aggregation operation to the primary server.'
-            server = ReadPreference.get(mode: :primary).select_server(cluster)
+            server = ServerSelector.get(mode: :primary).select_server(cluster)
             initial_query_op.execute(server.context)
           end
         end
