+++ lib/mongo/server_selector/primary.rb
@@ -14,13 +14,13 @@
 
 module Mongo
 
-  module ReadPreference
+  module ServerSelector
 
-    # Encapsulates specifications for selecting near servers given a list
+    # Encapsulates specifications for selecting the primary server given a list
     #   of candidates.
     #
     # @since 2.0.0
-    class Nearest
+    class Primary
       include Selectable
 
       # Get the name of the server mode type.
@@ -28,65 +28,58 @@ module Mongo
       # @example Get the name of the server mode for this preference.
       #   preference.name
       #
-      # @return [ Symbol ] :nearest
+      # @return [ Symbol ] :primary
       #
       # @since 2.0.0
       def name
-        :nearest
+        :primary
       end
 
       # Whether the slaveOk bit should be set on wire protocol messages.
       #   I.e. whether the operation can be performed on a secondary server.
       #
-      # @return [ true ] true
+      # @return [ false ] false
       #
       # @since 2.0.0
       def slave_ok?
-        true
+        false
       end
 
-      # Whether tag sets are allowed to be defined for this read preference.
+      # Whether tag sets are allowed to be defined for this server preference.
       #
-      # @return [ true ] true
+      # @return [ false ] false
       #
       # @since 2.0.0
       def tags_allowed?
-        true
+        false
       end
 
-      # Convert this read preference definition into a format appropriate
+      # Convert this server preference definition into a format appropriate
       #   for a mongos server.
       #
-      # @example Convert this read preference definition into a format
+      # @example Convert this server preference definition into a format
       #   for mongos.
-      #   preference = Mongo::ReadPreference::Nearest.new
+      #   preference = Mongo::ServerSelector::Primary.new
       #   preference.to_mongos
       #
-      # @return [ Hash ] The read preference formatted for a mongos server.
+      # @return [ nil ] nil
       #
       # @since 2.0.0
       def to_mongos
-        preference = { :mode => 'nearest' }
-        preference.merge!({ :tags => tag_sets }) unless tag_sets.empty?
-        preference
+        nil
       end
 
-      # Select the near servers taking into account any defined tag sets and
-      #   local threshold between the nearest server and other servers.
+      # Select the primary server from a list of candidates.
       #
-      # @example Select nearest servers given a list of candidates.
-      #   preference = Mongo::Serverreference::Nearest.new
-      #   preference.select_server(cluster)
+      # @example Select the primary server given a list of candidates.
+      #   preference = Mongo::ServerSelector::Primary.new
+      #   preference.select([candidate_1, candidate_2])
       #
-      # @return [ Array ] The nearest servers from the list of candidates.
+      # @return [ Array ] The primary server from the list of candidates.
       #
       # @since 2.0.0
       def select(candidates)
-        if tag_sets.empty?
-          near_servers(candidates)
-        else
-          near_servers(match_tag_sets(candidates))
-        end
+        primary(candidates)
       end
     end
   end
