+++ lib/mongo/server_selector/primary_preferred.rb
@@ -14,13 +14,13 @@
 
 module Mongo
 
-  module ReadPreference
+  module ServerSelector
 
-    # Encapsulates specifications for selecting servers, with
-    #   secondaries preferred, given a list of candidates.
+    # Encapsulates specifications for selecting servers, with the
+    #   primary preferred, given a list of candidates.
     #
     # @since 2.0.0
-    class SecondaryPreferred
+    class PrimaryPreferred
       include Selectable
 
       # Get the name of the server mode type.
@@ -28,11 +28,11 @@ module Mongo
       # @example Get the name of the server mode for this preference.
       #   preference.name
       #
-      # @return [ Symbol ] :secondary_preferred
+      # @return [ Symbol ] :primary_preferred
       #
       # @since 2.0.0
       def name
-        :secondary_preferred
+        :primary_preferred
       end
 
       # Whether the slaveOk bit should be set on wire protocol messages.
@@ -45,7 +45,7 @@ module Mongo
         true
       end
 
-      # Whether tag sets are allowed to be defined for this read preference.
+      # Whether tag sets are allowed to be defined for this server preference.
       #
       # @return [ true ] true
       #
@@ -54,40 +54,37 @@ module Mongo
         true
       end
 
-      # Convert this read preference definition into a format appropriate
+      # Convert this server preference definition into a format appropriate
       #   for a mongos server.
-      # Note that the read preference is not sent to mongos as part of the query
-      #   selector if there are no tag sets, for maximum backwards compatibility.
       #
-      # @example Convert this read preference definition into a format
+      # @example Convert this server preference definition into a format
       #   for mongos.
-      #   preference = Mongo::ReadPreference::SecondaryPreferred.new
+      #   preference = Mongo::ServerSelector::PrimaryPreferred.new
       #   preference.to_mongos
       #
-      # @return [ Hash ] The read preference formatted for a mongos server.
+      # @return [ Hash ] The server preference formatted for a mongos server.
       #
       # @since 2.0.0
       def to_mongos
-        return nil if tag_sets.empty?
-        preference = { mode: 'secondaryPreferred' }
-        preference.merge!({ tags: tag_sets })
+        preference = { :mode => 'primaryPreferred' }
+        preference.merge!({ :tags => tag_sets }) unless tag_sets.empty?
         preference
       end
 
       # Select servers taking into account any defined tag sets and
-      #   local threshold, with secondaries.
+      #   local threshold, with the primary preferred.
       #
       # @example Select servers given a list of candidates,
-      #   with secondaries preferred.
-      #   preference = Mongo::ReadPreference::SecondaryPreferred.new
+      #   with the primary preferred.
+      #   preference = Mongo::ServerSelector::PrimaryPreferred.new
       #   preference.select([candidate_1, candidate_2])
       #
       # @return [ Array ] A list of servers matching tag sets and acceptable
-      #   latency with secondaries preferred.
+      #   latency with the primary preferred.
       #
       # @since 2.0.0
       def select(candidates)
-        near_servers(secondaries(candidates)) + primary(candidates)
+        primary(candidates) + near_servers(secondaries(candidates))
       end
     end
   end
