+++ spec/mongo/server_selector/nearest_spec.rb
@@ -1,10 +1,11 @@
 require 'spec_helper'
 
-describe Mongo::ReadPreference::PrimaryPreferred do
-  include_context 'read preference'
+describe Mongo::ServerSelector::Nearest do
+
+  include_context 'server selector'
 
   it_behaves_like 'a read preference mode' do
-    let(:name) { :primary_preferred }
+    let(:name) { :nearest }
     let(:slave_ok) { true }
   end
 
@@ -12,10 +13,9 @@ describe Mongo::ReadPreference::PrimaryPreferred do
 
   describe '#to_mongos' do
 
-    context 'tag sets not provided' do
-
+    context 'tag set not provided' do
       it 'returns a read preference formatted for mongos' do
-        expect(read_pref.to_mongos).to eq({ :mode => 'primaryPreferred' })
+        expect(read_pref.to_mongos).to eq({ :mode => 'nearest' })
       end
     end
 
@@ -24,7 +24,7 @@ describe Mongo::ReadPreference::PrimaryPreferred do
 
       it 'returns a read preference formatted for mongos' do
         expect(read_pref.to_mongos).to eq(
-          { :mode => 'primaryPreferred', :tags => tag_sets}
+          { :mode => 'nearest', :tags => tag_sets}
         )
       end
     end
@@ -40,11 +40,11 @@ describe Mongo::ReadPreference::PrimaryPreferred do
       end
     end
 
-    context 'single primary candidate' do
+    context 'single primary candidates' do
       let(:candidates) { [primary] }
 
       it 'returns an array with the primary' do
-        expect(read_pref.select(candidates)).to eq( [primary] )
+        expect(read_pref.select(candidates)).to eq([primary])
       end
     end
 
@@ -52,25 +52,23 @@ describe Mongo::ReadPreference::PrimaryPreferred do
       let(:candidates) { [secondary] }
 
       it 'returns an array with the secondary' do
-        expect(read_pref.select(candidates)).to eq( [secondary] )
+        expect(read_pref.select(candidates)).to eq([secondary])
       end
     end
 
     context 'primary and secondary candidates' do
       let(:candidates) { [primary, secondary] }
-      let(:expected) { [primary, secondary] }
 
-      it 'returns an array with the primary first, then secondary' do
-        expect(read_pref.select(candidates)).to eq(expected)
+      it 'returns an array with the primary and secondary' do
+        expect(read_pref.select(candidates)).to match_array([primary, secondary])
       end
     end
 
-    context 'secondary and primary candidates' do
-      let(:candidates) { [secondary, primary] }
-      let(:expected) { [primary, secondary] }
+    context 'multiple secondary candidates' do
+      let(:candidates) { [secondary, secondary] }
 
-      it 'returns an array with the primary first, then secondary' do
-        expect(read_pref.select(candidates)).to eq(expected)
+      it 'returns an array with the secondaries' do
+        expect(read_pref.select(candidates)).to match_array([secondary, secondary])
       end
     end
 
@@ -88,24 +86,16 @@ describe Mongo::ReadPreference::PrimaryPreferred do
         context 'primary' do
           let(:candidates) { [primary] }
 
-          it 'returns array with primary' do
-            expect(read_pref.select(candidates)).to eq([primary])
+          it 'returns an empty array' do
+            expect(read_pref.select(candidates)).to be_empty
           end
         end
 
-        context 'matching_primary' do
+        context 'matching primary' do
           let(:candidates) { [matching_primary] }
 
-          it 'returns array with matching primary' do
+          it 'returns an array with the primary' do
             expect(read_pref.select(candidates)).to eq([matching_primary])
-          end
-        end
-
-        context 'matching secondary' do
-          let(:candidates) { [matching_secondary] }
-
-          it 'returns array with matching secondary' do
-            expect(read_pref.select(candidates)).to eq([matching_secondary])
           end
         end
 
@@ -116,50 +106,57 @@ describe Mongo::ReadPreference::PrimaryPreferred do
             expect(read_pref.select(candidates)).to be_empty
           end
         end
+
+        context 'matching secondary' do
+          let(:candidates) { [matching_secondary] }
+
+          it 'returns an array with the matching secondary' do
+            expect(read_pref.select(candidates)).to eq([matching_secondary])
+          end
+        end
       end
 
-      context 'multiple candidates' do
+      context 'mtuliple candidates' do
 
-        context 'no matching secondaries' do
+        context 'no matching servers' do
           let(:candidates) { [primary, secondary, secondary] }
 
-          it 'returns an array with the primary' do
-            expect(read_pref.select(candidates)).to eq([primary])
+          it 'returns an empty array' do
+            expect(read_pref.select(candidates)).to be_empty
           end
         end
 
         context 'one matching primary' do
           let(:candidates) { [matching_primary, secondary, secondary] }
 
-          it 'returns an array of the matching secondary, then primary' do
+          it 'returns an array with the matching primary' do
             expect(read_pref.select(candidates)).to eq([matching_primary])
           end
         end
 
         context 'one matching secondary' do
           let(:candidates) { [primary, matching_secondary, secondary] }
-          let(:expected) { [primary, matching_secondary] }
 
-          it 'returns an array of the primary and matching secondary' do
-            expect(read_pref.select(candidates)).to eq(expected)
+          it 'returns an array with the matching secondary' do
+            expect(read_pref.select(candidates)).to eq([matching_secondary])
           end
         end
 
         context 'two matching secondaries' do
           let(:candidates) { [primary, matching_secondary, matching_secondary] }
-          let(:expected) { [primary, matching_secondary, matching_secondary] }
+          let(:expected) { [matching_secondary, matching_secondary] }
 
-          it 'returns an array of the primary and matching secondaries' do
+          it 'returns an array with the matching secondaries' do
             expect(read_pref.select(candidates)).to eq(expected)
           end
         end
 
-        context 'one matching primary, one matching secondary' do
-          let(:candidates) { [primary, matching_secondary, secondary] }
-          let(:expected) { [primary, matching_secondary] }
+        context 'one matching primary and one matching secondary' do
+          let(:candidates) { [matching_primary, matching_secondary, secondary] }
+          let(:expected) { [matching_primary, matching_secondary] }
 
-          it 'returns an array of the primary and matching secondary' do
-            expect(read_pref.select(candidates)).to eq(expected)
+          it 'returns an array with the matching primary and secondary' do
+            expect(read_pref.select(candidates)).to match_array(expected)
           end
         end
       end
@@ -185,7 +182,6 @@ describe Mongo::ReadPreference::PrimaryPreferred do
           it 'returns array with far primary' do
             expect(read_pref.select(candidates)).to eq([far_secondary])
           end
-
         end
       end
 
@@ -193,28 +189,28 @@ describe Mongo::ReadPreference::PrimaryPreferred do
 
         context 'local primary, local secondary' do
           let(:candidates) { [primary, secondary] }
-          let(:expected) { [primary, secondary] }
 
-          it 'returns an array of the primary and secondary' do
-            expect(read_pref.select(candidates)).to eq(expected)
+          it 'returns array with primary and secondary' do
+            expect(read_pref.select(candidates)).to match_array(
+              [primary, secondary]
+            )
           end
         end
 
         context 'local primary, far secondary' do
           let(:candidates) { [primary, far_secondary] }
-          let(:expected) { [primary, far_secondary] }
 
-          it 'returns an array of the primary and far secondary' do
-            expect(read_pref.select(candidates)).to eq(expected)
+          # @todo: is this right?
+          it 'returns array with local primary' do
+            expect(read_pref.select(candidates)).to eq([primary])
           end
         end
 
         context 'far primary, local secondary' do
           let(:candidates) { [far_primary, secondary] }
-          let(:expected) { [far_primary, secondary] }
 
-          it 'returns an array of the far primary and secondary' do
-            expect(read_pref.select(candidates)).to eq(expected)
+          it 'returns array with local secondary' do
+            expect(read_pref.select(candidates)).to eq([secondary])
           end
         end
 
@@ -222,28 +218,28 @@ describe Mongo::ReadPreference::PrimaryPreferred do
           let(:candidates) { [far_primary, far_secondary] }
           let(:expected) { [far_primary, far_secondary] }
 
-          it 'returns an array of the far primary and far secondary' do
-            expect(read_pref.select(candidates)).to eq(expected)
+          it 'returns array with both servers' do
+            expect(read_pref.select(candidates)).to match_array(expected)
           end
         end
 
         context 'two local servers, one far server' do
 
-          context 'local primary, local secondary, far secondary' do
+          context 'local primary, local secondary' do
             let(:candidates) { [primary, secondary, far_secondary] }
             let(:expected) { [primary, secondary] }
 
-            it 'returns an array of the primary and secondary' do
-              expect(read_pref.select(candidates)).to eq(expected)
+            it 'returns array with local primary and local secondary' do
+              expect(read_pref.select(candidates)).to match_array(expected)
             end
           end
 
-          context 'two local secondaries' do
+          context 'two near secondaries' do
             let(:candidates) { [far_primary, secondary, secondary] }
-            let(:expected) { [far_primary, secondary, secondary] }
+            let(:expected) { [secondary, secondary] }
 
-            it 'returns an array with primary then two secondaries' do
-              expect(read_pref.select(candidates)).to eq(expected)
+            it 'returns array with the two local secondaries' do
+              expect(read_pref.select(candidates)).to match_array(expected)
             end
           end
         end
@@ -251,4 +247,3 @@ describe Mongo::ReadPreference::PrimaryPreferred do
     end
   end
 end
-
